backend/requirements.txt
django
djangorestframework
pillow
channels
channels-redis
redis
psycopg2-binary
Celery
python-dotenv
